Move libcontainerd/types to daemon/internal/libcontainerd/types
