+++ daemon/container/state_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/docker/docker/api/types/container"
-	libcontainerdtypes "github.com/docker/docker/libcontainerd/types"
+	libcontainerdtypes "github.com/docker/docker/daemon/internal/libcontainerd/types"
 )
 
 type mockTask struct {
