+++ daemon/internal/libcontainerd/remote/client_linux.go
@@ -11,7 +11,7 @@ import (
 	"github.com/containerd/containerd/v2/core/containers"
 	"github.com/containerd/containerd/v2/pkg/cio"
 	"github.com/containerd/log"
-	libcontainerdtypes "github.com/docker/docker/libcontainerd/types"
+	libcontainerdtypes "github.com/docker/docker/daemon/internal/libcontainerd/types"
 	"github.com/moby/sys/user"
 	"github.com/opencontainers/runtime-spec/specs-go"
 )
