+++ daemon/resize_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/docker/docker/daemon/container"
-	"github.com/docker/docker/libcontainerd/types"
+	"github.com/docker/docker/daemon/internal/libcontainerd/types"
 	"gotest.tools/v3/assert"
 )
