+++ daemon/start_notlinux.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/container"
-	"github.com/docker/docker/libcontainerd/types"
+	"github.com/docker/docker/daemon/internal/libcontainerd/types"
 	"github.com/opencontainers/runtime-spec/specs-go"
 )
